# Fix CI test failures and improve cross-platform compatibility