--- a/examples/simple.rs
+++ b/examples/simple.rs
@@ -1,3 +1,7 @@
+//! Simple example demonstrating basic Vietnamese input processing
+//!
+//! This example shows the most basic usage of the vi-engine library.
+
 fn main() {
     let inputs = vec![vec!['v', 'i', 'e', 't', '5', '6'], vec!['n', 'a', 'm']];