--- a/tests/vni.rs
+++ b/tests/vni.rs
@@ -1,3 +1,7 @@
+//! Tests for VNI Vietnamese input method
+//!
+//! This module contains snapshot tests for the VNI input method processing.
+
 mod shared;
 
 fn snapshot_transform(lines: &str) -> String {